api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
